Terminando a fusão entre a parte gráfica e o código de conversão
